requirements.txt
torch>=1.12.0
torchvision>=0.13.0
opencv-python>=4.6.0
numpy>=1.21.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
tensorboard>=2.9.0
open3d>=0.15.0
Pillow>=9.0.0
tqdm>=4.64.0
pyyaml>=6.0
scipy>=1.8.0
einops>=0.6.0